# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  my-web-app_com_smart-logistic-hub_internal_ai_dto.AIEventResponse:
    properties:
      confidence_score:
        type: number
      created_at:
        type: string
      event_code:
        type: string
      event_type:
        type: string
      gate_id:
        type: string
      id:
        type: integer
      license_plate:
        type: string
      low_confidence:
        type: boolean
      matched_driver_id:
        type: integer
      matched_trip_id:
        type: integer
      timestamp:
        type: string
    type: object
  my-web-app_com_smart-logistic-hub_internal_ai_dto.CreateAIEventRequest:
    properties:
      confidence_score:
        maximum: 1
        minimum: 0
        type: number
      event_code:
        type: string
      event_type:
        enum:
        - INBOUND
        - OUTBOUND
        type: string
      gate_id:
        type: string
      license_plate:
        type: string
      timestamp:
        type: string
    required:
    - event_type
    - gate_id
    - license_plate
    type: object
  my-web-app_com_smart-logistic-hub_internal_ai_dto.PaginatedResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_ai_dto.AIEventResponse'
        type: array
      limit:
        type: integer
      skip:
        type: integer
      total:
        type: integer
    type: object
  my-web-app_com_smart-logistic-hub_internal_ai_dto.UpdateAIEventRequest:
    properties:
      confidence_score:
        type: number
      event_type:
        type: string
      gate_id:
        type: string
      license_plate:
        type: string
      timestamp:
        type: string
    type: object
  my-web-app_com_smart-logistic-hub_internal_billing_dto.BillingResponse:
    properties:
      amount_total:
        type: number
      billing_code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      currency:
        type: string
      id:
        type: integer
      order_code:
        type: string
      paid_at:
        type: string
      payer_email:
        type: string
      payer_name:
        type: string
      payer_phone:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
    type: object
  my-web-app_com_smart-logistic-hub_internal_billing_dto.CreateBillingRequest:
    properties:
      amount_total:
        type: number
      billing_code:
        type: string
      created_by:
        type: string
      currency:
        type: string
      order_code:
        type: string
      payer_info:
        $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.PayerInfo'
      payment_method:
        type: string
      payment_status:
        type: string
    required:
    - billing_code
    - order_code
    - payer_info
    type: object
  my-web-app_com_smart-logistic-hub_internal_billing_dto.PaginatedResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.BillingResponse'
        type: array
      limit:
        type: integer
      skip:
        type: integer
      total:
        type: integer
    type: object
  my-web-app_com_smart-logistic-hub_internal_billing_dto.PayerInfo:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - name
    - phone
    type: object
  my-web-app_com_smart-logistic-hub_internal_billing_dto.UpdateBillingRequest:
    properties:
      paid_at:
        type: string
      payment_status:
        type: string
      transaction_id:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Smart Logistics Hub backend API. Authentication uses the "Authorization"
    header with format "Bearer &lt;token&gt;" (JWT issued by Keycloak).
  title: Smart Logistics Hub API
  version: 1.0.0
paths:
  /ai-events:
    get:
      parameters:
      - description: Filter by license plate
        in: query
        name: license_plate
        type: string
      - description: Filter by gate ID
        in: query
        name: gate_id
        type: string
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - default: 0
        description: Number of items to skip
        in: query
        name: skip
        type: integer
      - default: 20
        description: Max items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_ai_dto.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List AI events
      tags:
      - AI Events
    post:
      consumes:
      - application/json
      parameters:
      - description: AI event payload
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_ai_dto.CreateAIEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_ai_dto.AIEventResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an AI event
      tags:
      - AI Events
  /ai-events/{id}:
    delete:
      parameters:
      - description: AI event ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an AI event
      tags:
      - AI Events
    get:
      parameters:
      - description: AI event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_ai_dto.AIEventResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get an AI event by ID
      tags:
      - AI Events
    patch:
      consumes:
      - application/json
      parameters:
      - description: AI event ID
        in: path
        name: id
        required: true
        type: integer
      - description: AI event update
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_ai_dto.UpdateAIEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_ai_dto.AIEventResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an AI event
      tags:
      - AI Events
  /billing:
    get:
      parameters:
      - default: 0
        description: Number of items to skip
        in: query
        name: skip
        type: integer
      - default: 20
        description: Max items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List billing records
      tags:
      - Billing
    post:
      consumes:
      - application/json
      parameters:
      - description: Billing payload
        in: body
        name: billing
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.CreateBillingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.BillingResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a billing record
      tags:
      - Billing
  /billing/{id}:
    delete:
      parameters:
      - description: Billing ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a billing record
      tags:
      - Billing
    get:
      parameters:
      - description: Billing ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.BillingResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a billing record by ID
      tags:
      - Billing
    patch:
      consumes:
      - application/json
      parameters:
      - description: Billing ID
        in: path
        name: id
        required: true
        type: integer
      - description: Billing update
        in: body
        name: billing
        required: true
        schema:
          $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.UpdateBillingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.BillingResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a billing record
      tags:
      - Billing
  /billing/code/{billing_code}:
    get:
      parameters:
      - description: Billing code
        in: path
        name: billing_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.BillingResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a billing record by billing code
      tags:
      - Billing
  /billing/order/{order_code}:
    get:
      parameters:
      - description: Order code
        in: path
        name: order_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-web-app_com_smart-logistic-hub_internal_billing_dto.BillingResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a billing record by order code
      tags:
      - Billing
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"